Change readme to recommend loading font from a CDN.
The readme previously recommended loading the Inter font from
fonts.xyz.style, which sets `max-age=0` in the Cache-Control header.
This causes the font to be reloaded from the server on every page hit,
which results in an ugly Flash of Unstyled Text (FOUT) in Chrome.

By loading the font from the jsdelivr CDN, the FOUT is almost completely
eliminated because jsdelivr sets max-age to a high value and allows the
browser to cache the font.

Note that the new.css demo page at newcss.net uses jsdelivr for the
font.